+++ service/controller/app/resource/chart/delete.go
@@ -3,9 +3,9 @@ package chart
 import (
 	"context"
 
-	"github.com/giantswarm/app/v5/pkg/key"
+	"github.com/giantswarm/app/v6/pkg/key"
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/operatorkit/v5/pkg/resource/crud"
+	"github.com/giantswarm/operatorkit/v6/pkg/resource/crud"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
 	"github.com/giantswarm/app-operator/v5/service/controller/app/controllercontext"