+++ service/controller/catalog/resource/appcatalogentry/delete.go
@@ -3,7 +3,7 @@ package appcatalogentry
 import (
 	"context"
 
-	"github.com/giantswarm/app/v5/pkg/key"
+	"github.com/giantswarm/app/v6/pkg/key"
 	"github.com/giantswarm/microerror"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 )