+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Changed
 
+- Update Helm to `v3.6.3`.
+- Use controller-runtime client to remove CAPI dependency.
 - Use `apptestctl` to install CRDs in integration tests to avoid hitting GitHub rate limits.
 
 ### Removed